Refactor requirejs dependencies array
